change keymap name
